+++ frontend/src/features/pageContainer/pages/searchResultsPage/processSearch.js
@@ -1,0 +1,5 @@
+import store from '../../../../app/store';
+
+export const processSearch = () => {
+  return '';
+};
